Reduce stress test duration
Recently, I increased the stress test duration from 20 to 40 minutes
thinking "more is better" and in general, that is true. However, now
that we are running longer, it is becoming a regular occurrence that
stress test VMs in Cirrus are being pre-empted.

This commit puts us back to the previous 20 minute run times. At 20
minutes, we rarely got pre-empted.