Make some static bnf token & rule arrays const
This allows the compiler to put them in read only memory without additional analysis to find out that they are not written to.
To do this the related parameters of the `bnf_token_set` and `bnf_rule_set` functions are marked const, to declare that they won’t modify these arguments, as they don’t.