Fix soundness bug introduced in Pony 0.51.2 (#4283)
This commit includes a fix for the soundness bug introduced in 0.51.2 and  adds a set of simple "obviously-correct" tests of soundness / some completeness.

The bug was introduced because the change to safe_to_move in https://github.com/ponylang/ponyc/pull/4124/commits/c5113e74cd96bf0fb528a301b9e45534c3a63494 was correct from an isolation standpoint, but not from the standpoint of checking mutability, which is what the operator code was using it for. I.e. the change was valid only for the case of auto-recovering a constructor, and not mutating an existing object.

Now, mutability is checked before checking safety. This allows a better error message for the user in addition to fixing the bug.

Fixes #4282