Update dtrace probes
A user opened #4296 and noted that the probes were out of date. The
names of @params in comments need to match the names in the function
definitions.

While exploring the issue, I found that some functions lacked param
definitions, some had missing params, some had misnamed params, and some
functions were lacking useful information such as the actor in being run
where only the scheduler was present and the actor is the most useful
bit of information.

This commit fixes all the issues I found and in the process closes the
original issue.

Fixes #4296