Fix LLVM configuration override
With the introduction of CMake 3.21 and more recent than LLVM 12
versions of LLVM, the way we were setting configuration for our
LLVM build stopped working.

Matthias and I found this because, programs like ponydoc would no
longer link as libz was required despite our configuration indicating
otherwise.

After a lot of irritating testing and a considerable amount of rebuilding
of LLVM, we found the correct way to address for libz. This PR applies
that change to all the variables we are setting as we believe that this
approach needs to be taken for all of them.

Additionally, based on some CMake documentation, we believe this approach
only works with CMake 3.21 and later so as part of this PR we are moving
our required CMake from 3.11 to 3.21.

The CMake change required a number of CI container images to be updated to
install a version of CMake that meets our new requirements.