Always include llvm_args in the pass_opt_t struct
otherwise, when linking against libponyc from pony we cannot reliably use the fields coming after that,
as in ffi we don't know if the field is present or not and we cannot conditionally exclude or include a field in a struct in pony.

When compiled in release mode, this field will never be used and will remain a NULL pointer.