Remove explicit --platform from end user ponyc Dockerfiles
Previously, before the latest version of ponyup, when running on Glibc
versions of Linux, we could end up installing the wrong ponyc. To correct
for this, our Dockerfiles were supplying an explicit platform to use that
matched the distribution that formed the image base. This was required for
our "gnu" aka "glibc" image but not for musl. We only did it for musl to
provide symmetry.

The new version of ponyup will use lsb_release to correctly identify the
Glibc distribution and install the correct ponyc.