Stop building the "x86-64-unknown-linux-gnu" ponyc package (#4312)
The "x86-64-unknown-linux-gnu" package was intended to be "generic", however, it is not actually generic. It is a glibc version that is built on "some glibc using Linux" and will work if it is installed on a Linux that is compatible at a library level with the system that we built the package on.

Quite some time ago, Theo and I (who made the initial decision to include a "x86-64-unknown-linux-gnu" package) realized that the
 "x86-64-unknown-linux-gnu" package idea was "a very bad idea".

There are a couple large problems with this approach:

1- anyone using the package on glibc distro that isn't library compatible with our "x86-64-unknown-linux-gnu" build will have the installed ponyc fail in interesting and confounding ways. That's a bad user experience.

2- in order to not fall horribly behind, from time to time, we need to "rev" the environment used to build our "x86-64-unknown-linux-gnu" packages. For us that has meant changing the Ubuntu version periodically. That then means that people who had an environment that worked will do an update and install a ponyc that doesn't work because it is for a different set of libraries. That's a bad user experience as well.

This PR drops "x86-64-unknown-linux-gnu" from being built and mirrors a change in ponyup that dropped support for using the package via ponyup. See https://github.com/ponylang/ponyup/pull/245 for more information about the ponyup change.