Refactor `pony_type_t` for better struct packing.
This change moves the recently introduced field
`bool might_reference_actor` to a later part of `pony_type_t`
to improve the struct packing for the field layout,
by moving it into the "hole" left by a 32-bit field
that is situated between two pointer-sized fields.

This reduces the overall size of `pony_type_t` from 128 bytes
down to 120 bytes (the same size it was recently, before
the introduction of the `bool might_reference_actor` field).