LLVM 15 (#4327)
Updates to LLVM 15 and in the process removes the MergeRealloc optimization pass.

This optimization pass is not safe, and so it had to be removed.

This pass tries to remove calls to `pony_realloc` by making the original
pointer bigger to match the later realloc size, but it doesn't do any analysis
to prove that the reallocation didn't have some semantic importance.
For example, when using `Array.chop` or `String.chop` followed by a `push`
onto the left side of the chop point, the left side must be reallocated
in order to avoid mutating the right side. But this pass doesn't do any
analysis to check if the pointer may possibly be accessible from some SSA,
so it doesn't know that removing the reallocation will break the semantics.