Add an actual member to stacks (#4333)
These data structures in the Pony runtime act as a macro-named encapsulation layer
over an internal struct. This approach lets you cast from the macro-named struct pointer
to/from the internal struct pointer, pointing to the same memory address.

Prior to this commit, the stack data structure was using an empty struct for the
declaration of the macro-named struct. This is a lie (it doesn't reflect the reality
that the real content is non-empty) and it's inconsistent with how other data
structures are implemented in the other headers.

Overall, that's bad for maintainability and makes it easier to make a mistake
if anyone ever tries to dereference the macro-named pointer.