Add option to use pool implemenation which relies on memalign
This commit add an option to build pony with the pool implementation
based on posix_memalign instead of the default pool implementation.
This, combined with address/undefined behavior/etc sanitizers, allows
for easier debugging of memory issues within the compiler and runtime.
This commit also includes two small bugfixes in the compiler that
were discovered as part of this:

* lexer.c fix in `normalise_string` to prevent issues when
  `lexer->buflen` == 0
* names.c fix in `names_typeparam` to prevent access to an AST node
  that was just replaced

and a fix to the `Makefile` for running tests with `use=` flags
and a fix to make sure the test runner gets rebuilt whenever ponyc
gets rebuilt