fix `names_typeparam` to prevent access to an AST node that was just replaced
This was flagged by either the memory sanitizer or the undefined behavior
sanitizer.

Prior to this commit, the AST node is replaced and then the call to `ast_get`
tries to access the old node that has already been freed.

This commit ensures that the call to `ast_get` will correctly access the
newly AST node.