Fix for running tests with `use=` flags enabled
Prior to this commit, running `make test` after running
`make configure use=use_flag` would fail because the `Makefile`
would look for the test binaries in the wrong directory (i.e. in
`build/debug` instead of `build/debug-runtimestats`).

This commit updates the `Makefile` to try and determine the `use`
output directory from the CMake configure generated
`cmake_install.cmake` file so that tests can be run successfully.
It will fall back to the current behavior and use the default
build directory if it the file doesn't exist or it is unable to
extract the directory path.