Make sure the test runner gets rebuilt whenever ponyc gets rebuilt (#4362)
Prior to this commit, a call to `make build` would not rebuild the
test `runner` whenever `ponyc` would get rebuilt. This could
cause some confusion, especially if `libponyrt` changed, as one
would expect `make build` to ensure the `runner` is built using the
latest version of the runtime and compiler.

This commit ensures that the runner will get rebuilt whenever
`ponyc` gets rebuilt by adding `ponyc` as an explicit dependency
to the `runner` build command.