Use pointer tagging to avoid clearing chunks at start of GC
As part of https://github.com/ponylang/ponyc/pull/4143, the logic was
changed to avoid clearing chunks at the start of GC (except for
sizeclass 0) by using a sentinel value in the `chunk->shallow` field.

This commit, changes the implementation to rely on pointer tagging of
the `chunk->m` field using the lowest bit to track whether a chunk
needs to be cleared or not. The sentinel in the `chunk->shallow` is
no longer needed/used allowing for this optimization to work for
all sizeclasses (including the previously excluded sizeclass 0). All
accesses to the `chunk->m` field now go through the utility function
`get_m` to clear out the tagged bit first.