Use distinct directory when multiple `use=` flags are enabled (#4365)
Prior to this commit, the output directory used when multiple
`use=` flags were enabled would be the directory for the last
`use=` flag enabled in source code order.

This commit changes it so that the output directory uses a distinct
name that concats all the `use=` flags enabled together (i.e. the
output directory will be `debug-valgrind-runtimestats` instead of
`debug-runtimestats`).