Add workflow to manually update the lib cache
We cache `lib` in order to speed up our builds as building all of LLVM
takes a long long time. Once an update to any of the files that would
invalidate the cached libs is done, until a new cached version is done
on "main", it won't apply to new PRs.

This workflow allows us to update (if needed) the lib cache for any
given branch. In general, this would usually only be used on `main`.