Improve string-message-unbench stress tester (#4375)
This update does two things:

- Exits on fatal errors

Previously fatal errors could be missed in the logs (we never look) or
would only be seen in `debug` mode. Both fatal errors will now exit the
program immediately with a non-zero exit code which will result in
noticeable failure.

- Make messages a bit more stressful

Previously every message was the string "42". We are now sending larger messages
that grow more and in the process allocate more memory (in a not particularly
efficient fashion) to stress the runtime "just a bit" more than previously.