Add `make ponyc` for building just ponyc.
Sometimes while working on a compiler bug, it's annoying to
rebuild lots of tests/runners/etc when just trying to make
a new version of the compiler to do some manual testing with it.

This PR adds `make ponyc` which only uses the `ponyc` CMake target,
rather than the `all` target. Everything else in the invocation is
the same as it is for `make build`.