Move Windows CI from CirrusCI to GitHub Actions
CirrusCI no longer offers unlimited free CI to open source projects
so, we need to move as we use a lot of CI resources and quickly
exhaust what we get from Cirrus for free.

Due to some complications with moving over the existing CirrusCI
workflow, this is a "partial move". In Cirrus, we were using Windows
Containers for all CI. We can't do that in GitHub Actions as Windows
Containers aren't supported.

Getting an environment set up that supported LLDB has been difficult
so, as a stepping stone, this commit gets us set up with running
tests in LLDB.