Delete untagged stdlib builder images (#4408)
This commit adds a steps that runs after each update for the
tagged standard library builder images. The update will delete
all untagged builders.

Based on how we use the builder (build, push, use) we shouldn't ever
need any of the untagged builders. They are noise for us in the UI
of things that will never be used and waste storage for GitHub.