Move Windows CI from Cirrus to GitHub Actions
As part of this commit, the TCP tests were updated to use a different
port for each test as Windows in GitHub Actions was failing with
weird errors that seemed to indicate that ports where being kept open
longer than they should that resulted in weird errors. Running each
test on its own guaranteed port fixed the issues.