Fix failing CI tests that were run in LLDB 14 not reporting failures (#4417)
LLDB 14 and 15 (perhaps 16, we didn't test) has some issues.

It doesn't run "on crash" scripts if the crash occurred in any thread
but "the main thread". Additionally, it doesn't give us the back up
exit status message we would search for.

The result? CI tests run in LLDB 14 or 15 would fail but the runner
would report them as having passed.

This patch is from Gordon Tisher.