Update supported Alpine version to 3.17 and later (#4407)
With Alpine 3.17, Alpine stopped supporting libexecinfo. In the process,
the backtrace functionality was moved into musl directly. This makes it
very hard to support both pre and post 3.17 Alpine with the compiler.

We decided to not support 3.16 and earlier as it is a decent amount of
work to support both and 3.16 is EOL in about 9 months.

As part of the change, we are updating our public docker images to be
compatiable with the change. Our public Alpine images are now based off
of Alpine 3.18.