Avoid hangs in async pony_check properties when using actions
It was possible for calls from stray actors from previous executions of the 
same property to call into the current execution of the property 
(in `_PropertyRunner`) and thus confuse  the current run.

This is fixed now. Each property run, and also the property executions during 
shrinking, now is uniquely identified, and all calls using actions now need to
provide those unique ids. Inside the `_PropertyRunner` we now check if the call
we receive is from the current expected run or from a previous one, and discard
the call in the latter case.

There is no user-facing API change for the `pony_check` package coming from
this.