Move full-program test runner up one directory level (#4430)
This PR moves the full-program test runner code up one directory level so it is not mixed up with the actual test directories.  I was getting tired of hunting for it amid all the tests.

The test runner directory is now `test/full-program-runner`; the tests themselves now live in `test/full-program-tests`.