Fix cmake dependency for moved test runner
When I moved and renamed the test runner, I neglected to update the cmake target executable name.  This means the runner does in fact get built the first time you run `make`, but it won't get incrementally updated if sources change.

This PR updates the name of the executable so cmake can correctly track it.