Update PosixDate.format doc string (#4450)
Note the "not at all obvious" caveat where we normalize %p and %P
handling. Some locales will return an empty string for those format
strings which if they are the only thing in the format string makes
it impossible to know if there was success or failure from the format
so we short-circuit those and make all locales return empty string
if either formatting option is the only one in the format string.