Fix compiler bug that allows an unsafe data access pattern
Last November, Gordon identified that the compiler was allowing code
to recover a `val` from a `box` field. This is unsafe and very "doh".

After investigation, it was determined that this bug has existed since
Sylvan introduced viewpoint adaptation to the compiler almost a decade
ago.

I came up with a fix, but after talking with Sylvan, we changed to this
fix as it is the "proper pure theory fix".

Fixes #4244