Only include `ponyc.exe` in the `bin` directory of Windows release zip files
Previously the tests and benchmark executables were included, which doubled the size of the zip file.  This PR removes those files before zipping the release archive.

We do it this way because the Powershell command `Compress-Archive` cannot add a single file to a directory other than the root directory of the zip file, so we can't specify just `bin\ponyc.exe` in the list of things to add to the archive; it'll add it to the root directory.  Ponyc expects to find its lib and packages directories one directory below where it is.

Fixes #4272