Add Fedora 39 builder
The builder will be used to build nightly and release versions of ponyc
targetting Fedora 39.

No release notes are being added at this time, nor is RELEASE_PROCESS.md
being updated as until a nightly runs, we can't be sure this will fully work.

The docker image created for the builder can almost certainly be trimmed
down in size, but I am not a Fedora/DNF expert. I am looking to the requester
of Fedora 39 support to assist with trimming the image down.