Fix generation of invalid LLVM IR  (#4506)
Fix #4475

Before this change, `LLVM` reported an error during module verification when trying to compile the code referenced in issue #4475:

```pony
class Foo
  new create(a: U32) ? =>
    error

actor Main
  new create(env: Env) =>
    try
      let f = Foo(1)?
    end
```
This bug was caused by the failure to check for the presence of terminator instructions when generating `ret` instructions. So now, with this change, there's a check for the absence of a terminator instruction each time a `ret` instruction is generated.