Make heap small chunk size setting logic more precise/correct
Prior to this commit, the heap small chunk size setting logic would
bitwise-or in the size into the appropriate bits of the small chunk
but this is technically incorrect as it doesn't clear out the bits
first.

This commit changes the logic to ensure the relevant bits are cleared
first before the new size is bitwise-or'd in.