Disable actor heap large chunk recycling
The implementation of actor heap large chunk recycling from #4531
too naive and results in actors wasting huge amounts of time related
to large chunk recycling.

This commit effectively disables the large chunk recycling but
doesn't undo the code changes at the moment because the expectation
is that large chunk recycling will be re-enabled in the near future
with an improved implementation.