Fix bug in ASIO shutdown
Prior to this commit, when shutting down the ASIO subsystem, we
ignored the result of our waiting to join on the ASIO thread's
termination.

By ignoring the result, in cases where the join failed, we would
continue doing ASIO shutdown related activities while the ASIO
system was still running.

This was "not good".